backend/requirements.txt
Flask>=2.0
gunicorn>=20.0
python-dotenv>=0.21
flask-cors>=3.0
Flask-SQLAlchemy>=3.0
PyJWT>=2.0
passlib[bcrypt]>=1.7
marshmallow>=3.0
flask-marshmallow>=0.14
pytest>=7.0